Added check for an empty file.
In the case when no file is found, 404 will be returned.

+++ fetch/sec-metadata/resources/get-header.py
@@ -1,19 +1,24 @@
 import os
+import time
 
 def main(request, response):
-  path = "fetch/sec-metadata/resources/tmp"
+  path = "fetch/sec-metadata/resources/tmp/"
 
   ## Get the query parameter (filename) from URL ##
   filename =request.url_parts.query.split("=")[1]
 
-  ## Read and remove the file from tmp/ directory ##
-  file = open(path + filename, "r")
-  header = file.read()
-  file.close()
-  os.remove(path+filename)
-
-  ## Write the file contents to the response object ##
-  response.writer.write_status(200)
-  response.writer.end_headers()
-  response.writer.write(str(header))
-  response.close_connection = True
+  if not os.path.isfile(path + filename):
+    response.writer.write_status(404)
+    response.writer.end_headers()
+    response.close_connection = True
+  else:
+    ## Read and remove the file from tmp/ directory ##
+    file = open(path + filename, "r")
+    header = file.read()
+    file.close()
+    os.remove(path+filename)
+    ## Write the file contents to the response object ##
+    response.writer.write_status(200)
+    response.writer.end_headers()
+    response.writer.write(str(header))
+    response.close_connection = True
